Updated Nav bar to show only appropriate links
