+++ client/package.json
@@ -14,6 +14,7 @@
     "@mui/material": "^5.13.7",
     "@stripe/stripe-js": "^1.3.1",
     "animate.css": "^4.1.1",
+    "body-scroll-lock": "^4.0.0-beta.0",
     "gh-pages": "^5.0.0",
     "graphql": "^15.4.0",
     "gsap-trial": "^3.12.1",
