+++ client/src/components/Cart/style.css
@@ -7,22 +7,22 @@
   max-height: 60%;
   background-color: var(--light);
   overflow: auto;
-  padding: .5rem;
-  box-shadow: 0 0 1rem rgba(0, 0, 0, .5);
-  border-bottom-left-radius: .5rem;
+  padding: 0.5rem;
+  box-shadow: 0 0 1rem rgba(0, 0, 0, 0.5);
+  border-bottom-left-radius: 0.5rem;
 }
 
 .cart h2 {
   font-size: 1.5rem;
   border-bottom: 1px solid var(--dark);
-  padding-bottom: .5rem;
+  padding-bottom: 0.5rem;
   margin: 1rem 0;
 }
 
 .cart .close {
   position: absolute;
-  top: .5rem;
-  right: .5rem;
+  top: 0.5rem;
+  right: 0.5rem;
   cursor: pointer;
   color: var(--tertiary);
 }
@@ -33,13 +33,13 @@
 
 .cart-closed {
   position: fixed;
-  top: 2%;
+  top: 8%;
   right: 1%;
   font-size: 2rem;
   cursor: pointer;
   background-color: var(--secondary);
   border-radius: 50%;
-  padding: .25rem;
+  padding: 0.25rem;
   width: 50px;
   height: 50px;
 }
@@ -56,7 +56,7 @@
 .cart input {
   width: 50px;
   padding: 0;
-  margin: 0 .5rem 1rem 0;
+  margin: 0 0.5rem 1rem 0;
 }
 
 @media screen and (max-width: 768px) {
