+++ client/src/components/CategoryMenu/index.js
@@ -43,14 +43,12 @@ function CategoryMenu() {
 
   return (
     <div>
-      <h2>Choose a Category:</h2>
       {categories.map((item) => (
         <button
           key={item._id}
           onClick={() => {
             handleClick(item._id);
-          }}
-        >
+          }}>
           {item.name}
         </button>
       ))}
