+++ client/src/index.css
@@ -222,3 +222,24 @@ input:focus {
     width: 100%;
   }
 }
+
+.App {
+  text-align: center;
+}
+
+.modalDiv {
+  width: 100vw;
+  height: 100vh;
+  position: absolute;
+  top: 0;
+  background-color: rgba(91, 112, 131, 0.4);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.modal {
+  width: 350px;
+  height: 200px;
+  background-color: white;
+  border-radius: 5px;
+}
