+++ client/src/pages/Home.js
@@ -1,14 +1,12 @@
-import React from "react";
-import ProductList from "../components/ProductList";
-import CategoryMenu from "../components/CategoryMenu";
-import Cart from "../components/Cart";
+import React from 'react';
+import ProductList from '../components/ProductList';
+import CategoryMenu from '../components/CategoryMenu';
 
 const Home = () => {
   return (
     <div className="container">
       <CategoryMenu />
       <ProductList />
-      <Cart />
     </div>
   );
 };
